requirements.txt
numpy==1.25.2
pandas==2.1.4
PyMCubes==0.1.4
pysdf==0.1.9
scikit_learn==1.3.0
setuptools==59.6.0
torch==2.0.1
torch_ema==0.3
tqdm==4.66.1
trimesh==4.0.8
